% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accurate-glm.R
\docType{class}
\name{AccurateGLM-class}
\alias{AccurateGLM-class}
\title{S4 class for fitted AGLM}
\description{
S4 class for fitted AGLM
}
\section{Slots}{

\describe{
\item{\code{backend_model}}{A model object to be passed to backend API internally. Currently glmnet is used as backend,
and this slot holding a glmnet object.}

\item{\code{vars_info}}{A list of lists. Each element represents one predictor and contains information
of it and how to create its matrix representation from \code{data} slot.}
}}

