% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-aglm-input.R
\name{createAglmInput}
\alias{createAglmInput}
\title{Create aglm.input object}
\usage{
createAglmInput(x = NULL, x_UD = NULL, UD_vars = NULL)
}
\arguments{
\item{x}{input matrix or data.frame of size (nobs, _). Each row shold be an integer or numeric vector.}

\item{x_UD}{additional and optional input matrix of size (nobs, _), whose columns are replaced with U-dummy
matrices before fitting. Each row of \code{x_UD} should be an integer, character, or factor vector.}

\item{UD_vars}{integer or character optional vector. If an integer vector is given, this function replaces
columns whose indices are in \code{UD_vars} with U-dummy matrices before fitting. If a character vector is given,
this function replaces each column whose name is in \code{UD_vars} with U-dummy matrices before fitting.
\code{UD_vars} is ignored when \code{x_UD} is not NULL.
If both \code{UD_vars} and \code{x_UD} are not given (NULL values), columns with character types or factor types are
automatically handled as in \code{UD_vars}.}
}
\value{
a list with two elements \code{x_all} and \code{vars_info}
\itemize{
\item \code{x_all}: a data.frame which hold the whole data of x and x_UD
\item \code{vars_info}: a list of length dim(x_all)\link{2}. \code{vars_info[[i]]} represents some information of \code{x_all[, i]}.
}
}
\description{
Create aglm.input object
}
