% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{createEqualFreqBins}
\alias{createEqualFreqBins}
\title{Create bins by Equal Freq Binning}
\usage{
createEqualFreqBins(x_vec, nbin.max)
}
\arguments{
\item{x_vec}{A reference integer or numeric or ordered vector to be binned.}

\item{nbin.max}{An integer value which indicates the maximum counts of bins.
Note that this function makes \code{min(nbin.max, length(x_vec))} counts of bins.}
}
\value{
a numeric vector which indicates the boundaries of bins, with (number of bins + 1) elements.
}
\description{
Create bins by Equal Freq Binning
}
