% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{createEqualWidthBins}
\alias{createEqualWidthBins}
\title{Create bins by Equal Width Binning}
\usage{
createEqualWidthBins(left, right, nbin)
}
\arguments{
\item{left}{left value of the original interval.}

\item{right}{right value of the original interval.}

\item{nbin}{number of bins to be created.}
}
\value{
a numeric vector which indicates the boundaries of bins, with \code{nbin} elements.
}
\description{
Create bins by Equal Width Binning
}
