% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{createEvenBins}
\alias{createEvenBins}
\title{Create evenly cut bins.}
\usage{
createEvenBins(left, right, nbin)
}
\arguments{
\item{left}{left value of the original interval.}

\item{right}{right value of the original interval.}

\item{nbin}{number of bins to be created.}
}
\value{
a numeric vector which indicates the boundaries of bins, with (number of bins + 1) elements.
}
\description{
Create evenly cut bins.
}
