% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance-aglm.R
\name{deviance.AccurateGLM}
\alias{deviance.AccurateGLM}
\title{Extract the deviance from an AccurateGLM object}
\usage{
\method{deviance}{AccurateGLM}(fitted, ...)
}
\description{
Extract the deviance from an AccurateGLM object
}
