% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{executeBinning}
\alias{executeBinning}
\title{Execute binning for numerical data.}
\usage{
executeBinning(x_vec, breaks = NULL, nbin = 20, allow_na = FALSE)
}
\arguments{
\item{x_vec}{an integer or numeric vector, to be binned.}

\item{breaks}{a numeric vector which indicates the boundaries of bins, of length (number of bins + 1).
If NULL is set, evenly cut bins are generated and used.}

\item{nbin}{a number of bins which is used. Only used when \code{breaks} is not set.}

\item{allow_na}{a boolean value to tell whether all the elements should be binned or not (default is FALSE).
If set TRUE, the leftmost and rightmost boundaries of bins are enhanced to -Inf and +Inf, respectively.}
}
\value{
a list which has two members \code{labels} and \code{breaks}.
\itemize{
\item \code{labels}: an integer vector of \code{length(x_vec)}.
\code{(labels[i]==k)} indicates the i-th element of x_vec is in the k-th bin.
\item \code{breaks}: a numeric vector which indicates the boundaries of bins, of length (number of bins + 1).
This vector can be different from the \code{breaks} argument when \code{breaks} is not set or \code{allow_na} is FALSE.
}
}
\description{
Execute binning for numerical data.
}
