% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred-vars.R
\name{getDesignMatrix}
\alias{getDesignMatrix}
\title{Get design-matrix representation of PredVars objects}
\usage{
getDesignMatrix(preds, standardize_quantitative_vars = TRUE)
}
\arguments{
\item{preds}{A PredVars object}

\item{standardize_quantitative_vars}{A boolean value indicating quantitative values should be standardized.
Note that this option does not affect creations of dummy values (both O-dummies and U-dummies).}
}
\value{
A data.frame which represents the matrix representation of \code{preds}.
}
\description{
Get design-matrix representation of PredVars objects
}
