% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dummies.R
\name{getLVarMatForOneVec}
\alias{getLVarMatForOneVec}
\title{Create L-variable matrix for one variable}
\usage{
getLVarMatForOneVec(x_vec, breaks = NULL, nbin.max = 100, only_info = FALSE)
}
\arguments{
\item{x_vec}{A numeric vector representing original variable.}

\item{breaks}{A numeric vector representing breaks of bins (If \code{NULL}, automatically generated).}

\item{nbin.max}{The maximum number of bins (used only if \code{breaks=NULL}).}

\item{only_info}{If \code{TRUE}, only information fields of returned values are filled and no dummy matrix is returned.}
}
\value{
a list with the following fields:
\itemize{
\item \code{breaks}: Same as input
\item \code{dummy_mat}: The created L-variable matrix (only if \code{only_info=FALSE}).
}
}
\description{
Create L-variable matrix for one variable
}
\author{
Kenji Kondo
}
