% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dummies.R
\name{getODummyMatForOneVec}
\alias{getODummyMatForOneVec}
\title{Create a O-dummy matrix for one variable}
\usage{
getODummyMatForOneVec(
  x_vec,
  breaks = NULL,
  nbin.max = 100,
  only_info = FALSE,
  dummy_type = NULL
)
}
\arguments{
\item{x_vec}{A numeric vector representing original variable.}

\item{breaks}{A numeric vector representing breaks of bins (If \code{NULL}, automatically generated).}

\item{nbin.max}{The maximum number of bins (used only if \code{breaks=NULL}).}

\item{only_info}{If \code{TRUE}, only information fields of returned values are filled and no dummy matrix is returned.}

\item{dummy_type}{Used to control the shape of linear combinations obtained by O-dummies for quantitative variables (deprecated).}
}
\value{
a list with the following fields:
\itemize{
\item \code{breaks}: Same as input
\item \code{dummy_mat}: The created O-dummy matrix (only if \code{only_info=FALSE}).
}
}
\description{
Create a O-dummy matrix for one variable
}
\author{
Kenji Kondo
}
