% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-aglm.R
\name{print.AccurateGLM}
\alias{print.AccurateGLM}
\title{Print an AccurateGLM object}
\usage{
\method{print}{AccurateGLM}(fitted, digits = max(3, getOption("digits") -
  3), ...)
}
\description{
Print an AccurateGLM object
}
