% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accurate-glm.R
\docType{class}
\name{AccurateGLM-class}
\alias{AccurateGLM-class}
\title{S4 class for fitted AGLM}
\description{
S4 class for fitted AGLM
}
\section{Slots}{

\describe{
\item{\code{backend_models}}{Internally used model objects to be passed to backend functions.
Currently glmnet is used as a backend and this slot holding a glmnet object.}

\item{\code{vars_info}}{A list of list. Each element of \code{vars_info} represents one predictor variable and contains various informations of it.}
}}

