% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accurate-glm.R
\docType{class}
\name{AccurateGLM-class}
\alias{AccurateGLM-class}
\title{Class for results of \code{aglm()} and \code{cv.aglm()}}
\description{
Class for results of \code{aglm()} and \code{cv.aglm()}
}
\section{Slots}{

\describe{
\item{\code{backend_models}}{The fitted backend \code{glmnet} model is stored.}

\item{\code{vars_info}}{A list, each of whose element is information of one variable.}

\item{\code{lambda}}{Same as in the result of \link{cv.glmnet}.}

\item{\code{cvm}}{Same as in the result of \link{cv.glmnet}.}

\item{\code{cvsd}}{Same as in the result of \link{cv.glmnet}.}

\item{\code{cvup}}{Same as in the result of \link{cv.glmnet}.}

\item{\code{cvlo}}{Same as in the result of \link{cv.glmnet}.}

\item{\code{nzero}}{Same as in the result of \link{cv.glmnet}.}

\item{\code{name}}{Same as in the result of \link{cv.glmnet}.}

\item{\code{lambda.min}}{Same as in the result of \link{cv.glmnet}.}

\item{\code{lambda.1se}}{Same as in the result of \link{cv.glmnet}.}

\item{\code{fit.preval}}{Same as in the result of \link{cv.glmnet}.}

\item{\code{foldid}}{Same as in the result of \link{cv.glmnet}.}

\item{\code{call}}{An object of class \code{call}, corresponding to the function call when this \code{AccurateGLM} object is created.}

\item{\code{formula_info}}{A list to store information related to the formula object.}
}}

\author{
Kenji Kondo
}
