% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aglm.R
\name{aglm}
\alias{aglm}
\title{fit an AGLM model}
\usage{
aglm(x, y, x_UD = NULL, UD_vars = NULL, family = c("gaussian",
  "binomial", "poisson"), weights, offset = NULL, alpha = 1,
  nlambda = 100, lambda = NULL, standardize = TRUE,
  intercept = TRUE, thresh = 1e-07, exclude, lower.limits = -Inf,
  upper.limits = Inf, maxit = 1e+05, type.logistic = c("Newton",
  "modified.Newton"), standardize.response = FALSE)
}
\arguments{
\item{x}{input matrix or data.frame of size (nobs, _). Each row shold be an integer or numeric vector.}

\item{y}{response variable.}

\item{x_UD}{additional input matrix of size (nbos, _). optional. Each row of \code{x_UD} should be an integer,
character, or factor vector and this function replaces it with U-dummy matrix before fitting.}

\item{UD_vars}{integer or character vector. If an integer vector is set, this function replaces columns whose
indices are in \code{UD_vars} into U-dummy matrices before fitting. If a character vector is set, this function
replaces each column whose name is in \code{UD_vars} with U-dummy matrix before fitting.
Note that the columns of \code{x} are rearanged internally and could affect on outputs of print() and
plot() when \code{UD_vars} is set. \code{UD_vars} is ignored when \code{x_UD} is not NULL.
If both \code{x_UD} and \code{UD_vars} are NULL, columns with character types and factor types are automatically
replaced with U-dummy matrices before fitting. The columns of \code{x} are also rearanged internally in this case.}

\item{family}{response type. Currently gaussian, binomial, and poisson are supported.}

\item{standardize}{logical flag for x variable standardization, prior to fitting the model sequence.
Note that aglm() standardizes only x and never x_UD.}

\item{...}{other arguments are passed directly to glmnet() and have same default values.}
}
\value{
list with two elements \code{backend_result} and \code{vars_info}.
}
\description{
fit an AGLM model
}
