% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aglm.R
\name{aglm}
\alias{aglm}
\title{fit an AGLM model}
\usage{
aglm(x, x_UD, y, family = c("gaussian", "binomial", "poisson"), weights,
  offset = NULL, alpha = 1, nlambda = 100, lambda = NULL,
  standardize = TRUE, intercept = TRUE, thresh = 1e-07,
  dfmax = nvars + 1, exclude, lower.limits = -Inf,
  upper.limits = Inf, maxit = 1e+05, type.logistic = c("Newton",
  "modified.Newton"), standardize.response = FALSE)
}
\arguments{
\item{x}{input matrix or data.frame of size (nobs, _). Each row shold be an integer or numeric vector.}

\item{x_UD}{additional input matrix of size (nbos, _). Each row should be an integer or character or factor vector.}

\item{y}{response variable.}

\item{family}{response type. Currently gaussian, binomial, and poisson are supported.}

\item{standardize}{logical flag for x variable standardization, prior to fitting the model sequence.
Note that aglm() standardizes only x and never x_UD.}

\item{...}{other arguments are passed directly to glmnet() and have same default values.}
}
\value{
list with two elements \code{backend_result} and \code{vars_info}.
}
\description{
fit an AGLM model
}
