% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aglm.R
\name{aglm}
\alias{aglm}
\title{fit an AGLM model}
\usage{
aglm(x, y, qualitative_vars_UD_only = NULL,
  qualitative_vars_both = NULL, qualitative_vars_OD_only = NULL,
  quantitative_vars = NULL, add_linear_columns = TRUE,
  add_OD_columns_of_qualitatives = TRUE,
  add_intersection_columns = TRUE, standardize = TRUE,
  family = c("gaussian", "binomial", "poisson"), weights,
  offset = NULL, alpha = 1, nlambda = 100, lambda.min.ratio = NULL,
  lambda = NULL, intercept = TRUE, thresh = 1e-07, dfmax = NULL,
  pmax = NULL, exclude, penalty.factor = NULL, lower.limits = -Inf,
  upper.limits = Inf, maxit = 1e+05, type.gaussian = NULL,
  type.logistic = c("Newton", "modified.Newton"),
  standardize.response = FALSE)
}
\arguments{
\item{x}{An input matrix or data.frame to be fitted.}

\item{y}{An integer or numeric vector which represents response variable.}

\item{qualitative_vars_UD_only}{A list of indices or names for specifying which columns are qualitative and need only U-dummy representations.}

\item{qualitative_vars_both}{A list of indices or names for specifying which columns are qualitative and need both U-dummy and O-dummy representations.}

\item{qualitative_vars_OD_only}{A list of indices or names for specifying which columns are qualitative and need only O-dummy representations.}

\item{quantitative_vars}{A list of indices or names for specyfying which columns are quantitative.}

\item{add_linear_columns}{A boolean value which indicates whether this function uses linear effects or not.}

\item{add_OD_columns_of_qualitatives}{A boolean value which indicates whether this function use O-dummy representations for qualitative and ordinal variables or not.}

\item{add_intersection_columns}{A boolean value which indicates whether this function uses intersection effects or not.}

\item{family}{Response type. Currently "gaussian", "binomial", and "poisson" are supported.}

\item{...}{Other arguments are passed directly to backend (currently glmnet() is used), and if not given, backend API's default values are used to call backend functions.}
}
\value{
An AccurateGLM object, fitted to the data (x, y)
}
\description{
fit an AGLM model
}
