% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aglm.R
\name{aglm}
\alias{aglm}
\title{fit an AGLM model}
\usage{
aglm(x, y, x_UD = NULL, UD_vars = NULL, family = c("gaussian",
  "binomial", "poisson"), weights, offset = NULL, alpha = 1,
  nlambda = 100, lambda.min.ratio = NULL, lambda = NULL,
  standardize = TRUE, intercept = TRUE, thresh = 1e-07,
  dfmax = NULL, pmax = NULL, exclude, penalty.factor = NULL,
  lower.limits = -Inf, upper.limits = Inf, maxit = 1e+05,
  type.gaussian = NULL, type.logistic = c("Newton", "modified.Newton"),
  standardize.response = FALSE)
}
\arguments{
\item{x}{Explanatory or predictor variables used for fitting AGLM.
Although this function expects \code{x} would a PredVars object, it also can be matrix or data.frame.
Is such cases, this function automatically convert it to a PredVars object by calling
\code{newPredVars(x, x_UD, UD_vars)} and use that PredVars object as it is initially given.
Furthermore, \code{x} could be NULL in case where \code{x_UD} is given.
See descriptions of \code{newPredVars()} for details of non-PredVars cases.}

\item{y}{An integer or numeric vector which represents response variable.}

\item{x_UD, UD_vars}{See descriptions of \code{x} for details. Note that these values are required
only if a non-PredVar value are given as \code{x}.}

\item{family}{Response type. Currently "gaussian", "binomial", and "poisson" are supported.}

\item{standardize}{A boolean value which indicates whether standardizes each column of \code{x}.
Note that columns of \code{x_UD} and columns of \code{x} specified by \code{UD_vars} are not standardized
even if \code{standardize=TRUE}, because this function considers these columns as quantitative.}

\item{...}{Other arguments are passed directly to backend API (currently glmnet() is used),
and if not given, backend API's default values are used.}
}
\value{
An AccurateGLM object, fitted to the data (x, y)
}
\description{
fit an AGLM model
}
