% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aglm-input.R
\docType{class}
\name{AGLM_Input-class}
\alias{AGLM_Input-class}
\title{S4 class for input}
\description{
S4 class for input
}
\section{Slots}{

\describe{
\item{\code{vars_info}}{A list of list. Each element has some information of one feature.}

\item{\code{data}}{A data.frame which contains original data itself.}
}}

