% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef-aglm.R
\name{coef.AccurateGLM}
\alias{coef.AccurateGLM}
\title{Extract the deviance from an AccurateGLM object.}
\usage{
\method{coef}{AccurateGLM}(model, s = NULL, exact = FALSE, ...)
}
\arguments{
\item{model}{An AccurateGLM object.}

\item{s}{Value(s) of the penalty parameter \code{lambda} at which predictions are required.
Default is the entire sequence used to create the model.}

\item{...}{Other arguments are passed directly to \code{deviance} functions of \code{model@backend_models}.}
}
\description{
Extract the deviance from an AccurateGLM object.
}
