% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{createEqualFreqBins}
\alias{createEqualFreqBins}
\title{Create bins (equal frequency binning)}
\usage{
createEqualFreqBins(x_vec, nbin.max)
}
\arguments{
\item{x_vec}{A numeric vector, whose quantiles are used as breaks.}

\item{nbin.max}{The maximum number of bins.}
}
\value{
A numeric vector representing breaks obtained by binning.
Note that the number of bins is equal to \code{min(nbin.max, length(x_vec))}.
}
\description{
Create bins (equal frequency binning)
}
\author{
Kenji Kondo
}
