% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{createEqualWidthBins}
\alias{createEqualWidthBins}
\title{Create bins (equal width binning)}
\usage{
createEqualWidthBins(left, right, nbin)
}
\arguments{
\item{left}{The leftmost value of the interval to be binned.}

\item{right}{The rightmost value of the interval to be binned.}

\item{nbin}{The number of bins.}
}
\value{
A numeric vector representing breaks obtained by binning.
}
\description{
Create bins (equal width binning)
}
\author{
Kenji Kondo
}
