% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-aglm.R
\name{cv.aglm}
\alias{cv.aglm}
\title{Cross-validation for AGLM models}
\usage{
cv.aglm(x, y, qualitative_vars_UD_only = NULL,
  qualitative_vars_both = NULL, qualitative_vars_OD_only = NULL,
  quantitative_vars = NULL, add_linear_columns = TRUE,
  add_OD_columns_of_qualitatives = TRUE,
  add_interaction_columns = TRUE, OD_type_of_quantitatives = "C",
  bins_list = NULL, bins_names = NULL, weights, offset = NULL,
  lambda = NULL, type.measure, nfolds = 10, foldid, grouped = TRUE,
  keep = FALSE, parallel = FALSE, standardize = TRUE,
  family = c("gaussian", "binomial", "poisson"), alpha = 1,
  nlambda = 100, lambda.min.ratio = NULL, intercept = TRUE,
  thresh = 1e-07, dfmax = NULL, pmax = NULL, exclude,
  penalty.factor = NULL, lower.limits = -Inf, upper.limits = Inf,
  maxit = 1e+05, type.gaussian = NULL, type.logistic = c("Newton",
  "modified.Newton"), standardize.response = FALSE)
}
\arguments{
\item{x}{An input matrix or data.frame to be fitted.}

\item{y}{An integer or numeric vector which represents response variable.}

\item{qualitative_vars_UD_only}{A list of indices or names for specifying which columns are qualitative and need only U-dummy representations.}

\item{qualitative_vars_both}{A list of indices or names for specifying which columns are qualitative and need both U-dummy and O-dummy representations.}

\item{qualitative_vars_OD_only}{A list of indices or names for specifying which columns are qualitative and need only O-dummy representations.}

\item{quantitative_vars}{A list of indices or names for specyfying which columns are quantitative.}

\item{add_linear_columns}{A boolean value which indicates whether this function uses linear effects or not.}

\item{add_OD_columns_of_qualitatives}{A boolean value which indicates whether this function use O-dummy representations for qualitative and ordinal variables or not.}

\item{add_interaction_columns}{A boolean value which indicates whether this function uses interaction effects or not.}

\item{OD_type_of_quantitatives}{A character value which indicates how O-dummy matrices of quantitative
values are constructed. Choose 'C'(default) or 'J'.
\itemize{
\item 'C': Continuous-type dummies, which result continuous contribution curves.
\item 'J': Jum-type dummies, which result contribution curves with jumps.
}}

\item{bins_list}{A list of numeric vectors, each element of which is used as breaks when binning of a quantitative variable or a qualitative variable with order.}

\item{bins_names}{A list of column name or column index, each name or index of which specifies which column of \code{x} is binned used with an element of \code{bins_list} in the same position.}

\item{family}{Response type. Currently "gaussian", "binomial", and "poisson" are supported.}

\item{...}{Other arguments are passed directly to backend (currently cv.glmnet() is used), and if not given, backend API's default values are used to call backend functions.}
}
\value{
Result of cross-validation.
}
\description{
Cross-validation for AGLM models
}
