% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cva-aglm.R
\name{cva.aglm}
\alias{cva.aglm}
\title{Cross-validation for both alpha and lambda}
\usage{
cva.aglm(x, y, alpha = seq(0, 1, len = 11)^3, nfolds = 10,
  foldid = sample(rep(seq(nfolds), length = nrow(x))),
  parallel.alpha = FALSE, ...)
}
\arguments{
\item{x}{An input matrix or data.frame to be fitted.}

\item{y}{An integer or numeric vector which represents response variable.}

\item{alpha}{A numeric vector which represents alpha values to be examined.}

\item{nfolds}{The number of folds. The default value is 10.}

\item{foldid}{An integer vector of values between 1 and nfold identifying what fold each observation is in.}

\item{parallel.alpha}{not implemented yet.}

\item{...}{Other arguments are passed directly to cv.aglm().}
}
\value{
Result of cross-validation. See \code{CVA_AccurateGLM}'s document for more details.
}
\description{
Cross-validation for both alpha and lambda
}
