% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accurate-glm.R
\docType{class}
\name{CVA_AccurateGLM-class}
\alias{CVA_AccurateGLM-class}
\title{S4 class for the result of cva.aglm function}
\description{
S4 class for the result of cva.aglm function
}
\section{Slots}{

\describe{
\item{\code{models_list}}{Results of cv.glmnet() for all the values of alpha.}

\item{\code{alpha}}{A numeric values specifying alpha values to be examined.}

\item{\code{nfolds}}{An integer value specifying the number of folds.}

\item{\code{alpha.min}}{The alpha value which achieves the minimum loss.}

\item{\code{alpha.min.index}}{An integer value specifying the index of \code{alpha.min} in \code{alpha}.}

\item{\code{lambda.min}}{The lambda value which achieves the minimum loss, when combined with \code{alpha.min}.}
}}

