% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accurate-glm.R
\docType{class}
\name{CVA_AccurateGLM-class}
\alias{CVA_AccurateGLM-class}
\title{Class for results of \code{cva.aglm()}}
\description{
Class for results of \code{cva.aglm()}
}
\section{Slots}{

\describe{
\item{\code{models_list}}{A list consists of \code{cv.glmnet()}'s results for all \eqn{\alpha} values.}

\item{\code{alpha}}{Same as in \link{cv.aglm}.}

\item{\code{nfolds}}{Same as in \link{cv.aglm}.}

\item{\code{alpha.min.index}}{The index of \code{alpha.min} in the vector \code{alpha}.}

\item{\code{alpha.min}}{The \eqn{\alpha} value achieving the minimum loss among all the values of \code{alpha}.}

\item{\code{lambda.min}}{The \eqn{\lambda} value achieving the minimum loss when \eqn{\alpha} is equal to \code{alpha.min}.}

\item{\code{call}}{An object of class \code{call}, corresponding to the function call when this \code{CVA_AccurateGLM} object is created.}
}}

\author{
Kenji Kondo
}
