% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance-aglm.R
\name{deviance.AccurateGLM}
\alias{deviance.AccurateGLM}
\title{Extract the deviance from an AccurateGLM object.}
\usage{
\method{deviance}{AccurateGLM}(object, ...)
}
\arguments{
\item{object}{An AccurateGLM object.}

\item{...}{Other arguments are passed directly to \code{deviance} functions of \code{model@backend_models}.}
}
\description{
Extract the deviance from an AccurateGLM object.
}
