% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance-aglm.R
\name{deviance.AccurateGLM}
\alias{deviance.AccurateGLM}
\title{Get deviance}
\usage{
\method{deviance}{AccurateGLM}(object, ...)
}
\arguments{
\item{object}{A model object obtained from \code{aglm()} or \code{cv.aglm()}.}

\item{...}{Other arguments are passed directly to \code{deviance.glmnet()}.}
}
\value{
The value of deviance extracted from the object \code{object}.
}
\description{
Get deviance
}
\author{
Kenji Kondo
}
