% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{executeBinning}
\alias{executeBinning}
\title{Execute binning for numerical data.}
\usage{
executeBinning(x_vec, breaks = NULL, nbin.max = 100, method = "freq")
}
\arguments{
\item{x_vec}{an integer or numeric or ordered vector, to be binned.}

\item{breaks}{a numeric vector which indicates the boundaries of bins, of length (number of bins - 1).
If NULL is set, bins are generated using the method which is specified by the \code{method} argument.}

\item{nbin.max}{a maximum number of bins which is generated by this function. Only used when \code{breaks} is not set.}

\item{method}{used for specifying binning method. "freq": equal freq binning (default), "width": equal width binning.
Ignore if \code{breaks} is set.}
}
\value{
a list which has two members \code{labels} and \code{breaks}.
\itemize{
\item \code{labels}: an integer vector of \code{length(x_vec)}.
\code{(labels[i]==k)} indicates the i-th element of x_vec is in the k-th bin.
\item \code{breaks}: a numeric vector which indicates the boundaries of bins, of length (number of bins - 1).
}
}
\description{
Execute binning for numerical data.
}
