% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred-vars.R
\name{getDesignMatrix}
\alias{getDesignMatrix}
\title{Get design-matrix representation of PredVars objects}
\usage{
getDesignMatrix(preds)
}
\arguments{
\item{preds}{A PredVars object}
}
\value{
A data.frame which represents the matrix representation of \code{preds}.
}
\description{
Get design-matrix representation of PredVars objects
}
