% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aglm-input.R
\name{getDesignMatrix}
\alias{getDesignMatrix}
\title{Get design-matrix representation of AGLM_Input objects}
\usage{
getDesignMatrix(x)
}
\arguments{
\item{x}{An AGLM_Input object}
}
\value{
A data.frame which represents the matrix representation of \code{x}.
}
\description{
Get design-matrix representation of AGLM_Input objects
}
