% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dummies.R
\name{getODummyMatForOneVec}
\alias{getODummyMatForOneVec}
\title{Get O-dummy matrix for one-dimensional vector}
\usage{
getODummyMatForOneVec(
  x_vec,
  breaks = NULL,
  nbin.max = 100,
  only_info = FALSE,
  dummy_type = NULL
)
}
\arguments{
\item{x_vec}{An integer or numeric vector to be converted into dummy matrix.}

\item{breaks}{A numeric vector which indicates the boundaries of bins, of length (number of bins + 1).
If NULL, evenly cut bins are automatically generated and used.}

\item{nbin.max}{A maximum number of bins which is used. Only used when \code{breaks} is not set.}

\item{only_info}{A logical value. If TRUE, actual creation of dummy matrix is omitted.}

\item{dummy_type}{A character value. Choose \code{"C"}(default) or \code{"J"}. For integer or numeric \code{x_vec},
\code{dummy_type="C"} is used as default. Otherwise, \code{dummy_type="J"} is used as default.
\itemize{
\item \code{"C"}: Continuous-type dummies, which result continuous contribution curves.
\item \code{"J"}: Jump-type dummies, which result contribution curves with jumps.
}}
}
\value{
a list with two members \code{breaks} and \code{dummy_mat}.
\itemize{
\item \code{breaks}: Same as input
\item \code{dummy_mat}: An integer matrix with size (length of \code{x_vec}, length of \code{breaks} minus 1).
\code{dummy_mat[i, j]} is 1 if and only if \verb{breaks[i] < x_vec[i] <= breaks[i+1]}, and 0 otherwise.
Note that, in case where \code{x_vec[i]} is outside of \verb{(breaks[1], breaks[length(breaks)]]},
\code{x_vec[i]} is considered to be in the first bin if \code{x_vec[i] <= breaks[1]}, and
be in the last bin if \code{x_vec[i] > breaks[length(breaks)]}.
Omitted if \code{only_info=TRUE}
}
}
\description{
Get O-dummy matrix for one-dimensional vector
}
