% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dummies.R
\name{getUDummyMatForOneVec}
\alias{getUDummyMatForOneVec}
\title{Get U-dummy matrix for one-dimensional vector}
\usage{
getUDummyMatForOneVec(x_vec, drop_last = TRUE)
}
\arguments{
\item{x_vec}{non-numeric vector to be converted into dummy matrix.}
}
\value{
a list with two members \code{levels} and \code{dummy_mat}.
\itemize{
\item \code{levels}: a character vector. unique set of possible values. sorted.
\item \code{dummy_mat}: an integer matrix with size (length of \code{x_vec}, length of \code{levels} or minus 1 when \code{drop_last=TRUE}).
\code{dummy_mat[i, j]} is 1 if and only if \code{x_vec[i] == levels[j]}, and 0 otherwise.
}
}
\description{
Get U-dummy matrix for one-dimensional vector
}
