% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dummies.R
\name{getUDummyMatForOneVec}
\alias{getUDummyMatForOneVec}
\title{Get U-dummy matrix for one-dimensional vector}
\usage{
getUDummyMatForOneVec(
  x_vec,
  levels = NULL,
  drop_last = TRUE,
  only_info = FALSE
)
}
\arguments{
\item{x_vec}{A non-numeric vector to be converted into dummy matrix.}

\item{levels}{A character vector indicates unique set of possible values.
If NULL, all the unique values of \code{x_vec} are used.}

\item{drop_last}{A logical value. If TRUE, the last column of dummy matrix is dropped.}

\item{only_info}{A logical value. If TRUE, actual creation of dummy matrix is omitted.}
}
\value{
a list with two members \code{levels} and \code{dummy_mat}.
\itemize{
\item \code{levels}: Same as input
\item \code{drop_last}: Same as input
\item \code{dummy_mat}: An integer matrix with size (length of \code{x_vec}, length of \code{levels} or minus 1 when \code{drop_last=TRUE}).
\code{dummy_mat[i, j]} is 1 if and only if \code{x_vec[i] == levels[j]}, and 0 otherwise.
Omitted if \code{only_info=TRUE}
}
}
\description{
Get U-dummy matrix for one-dimensional vector
}
