% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-aglm.R
\name{plot.AccurateGLM}
\alias{plot.AccurateGLM}
\title{Plot coefficients from an AccurateGLM object}
\usage{
\method{plot}{AccurateGLM}(fitted, xvar = c("norm", "lambda", "dev"),
  label = FALSE, ...)
}
\description{
Plot coefficients from an AccurateGLM object
}
