% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-aglm.R
\name{predict.AccurateGLM}
\alias{predict.AccurateGLM}
\title{Make predictions for new data}
\usage{
\method{predict}{AccurateGLM}(
  object,
  newx = NULL,
  s = NULL,
  type = c("link", "response", "coefficients", "nonzero", "class"),
  exact = FALSE,
  newoffset,
  ...
)
}
\arguments{
\item{object}{A model object obtained from \code{aglm()} or \code{cv.aglm()}.}

\item{newx}{A design matrix for new data.
See the description of \code{x} in \link{aglm} for more details.}

\item{s}{Same as in \link{predict.glmnet}.}

\item{type}{Same as in \link{predict.glmnet}.}

\item{exact}{Same as in \link{predict.glmnet}.}

\item{newoffset}{Same as in \link{predict.glmnet}.}

\item{...}{Other arguments are passed directly when calling \code{predict.glmnet()}.}
}
\value{
The returned object depends on \code{type}.
See \link{predict.glmnet} for more details.
}
\description{
Make predictions for new data
}
\references{
Suguru Fujita, Toyoto Tanaka, Kenji Kondo and Hirokazu Iwasawa. (2020)
\emph{AGLM: A Hybrid Modeling Method of GLM and Data Science Techniques}, \cr
\url{https://www.institutdesactuaires.com/global/gene/link.php?doc_id=16273&fg=1} \cr
\emph{Actuarial Colloquium Paris 2020}
}
\author{
\itemize{
\item Kenji Kondo,
\item Kazuhisa Takahashi and Banno (worked on L-Variable related features)
}
}
