% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-aglm.R
\name{predict.AccurateGLM}
\alias{predict.AccurateGLM}
\title{Predicts response variables for new explanatory variables.}
\usage{
\method{predict}{AccurateGLM}(fitted, newx, newx_UD = NULL,
  UD_vars = NULL, standardize_quantitative_vars = TRUE, s = NULL,
  type = c("link", "response", "coefficients", "nonzero", "class"),
  exact = FALSE, newoffset, ...)
}
\arguments{
\item{fitted}{An AccurateGLM object fitted to data by aglm() function.}

\item{newx, newx_UD, UD_vars}{New data or predictor variables corresponding to response variables
to be predicted. See descriptions of aglm() for more details of usages.}

\item{standardize_quantitative_vars}{same as aglm()'s \code{standardize_quantitative_vars}.}

\item{...}{Other arguments are passed directly to backend (currently glmnet() is used), and
if not given, backend API's default values are used to call backend functions.}
}
\value{
same as glmnet()
}
\description{
Predicts response variables for new explanatory variables.
}
