% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-aglm.R
\name{predict.AccurateGLM}
\alias{predict.AccurateGLM}
\title{Make predictions from a fitted AccurateGLM}
\usage{
\method{predict}{AccurateGLM}(
  object,
  newx = NULL,
  s = NULL,
  type = c("link", "response", "coefficients", "nonzero", "class"),
  exact = FALSE,
  newoffset,
  ...
)
}
\arguments{
\item{object}{An AccurateGLM object.}

\item{newx}{An input matrix or data.frame used for predictions.}

\item{s}{Value(s) of the penalty parameter \code{lambda} at which predictions are required.
Default is the entire sequence used to create the model.}

\item{type}{Type of prediction required.
\itemize{
\item Type \code{"link"} gives the linear predictors for \code{"binomial"}, \code{"poisson"} models, and for \code{"gaussian"} models it gives the fitted values.
\item Type \code{"response"} gives the fitted probabilities for \code{"binomial"}, fitted mean for \code{"poisson"}, and for \code{"gaussian"} models it is equivalent to type \code{"link"}.
\item Type \code{"coefficients"} computes the coefficients at the requested values for \code{s}.
Note that for \code{"binomial"} models, results are returned only for the class corresponding to the second level of the factor response.
\item Type \code{"class"} applies only to \code{"binomial"}, and produces the  class label corresponding to the maximum probability.
\item Type \code{"nonzero"} returns a list of the indices of the nonzero coefficients for each value of \code{s}.
}}

\item{exact}{Same as predict.glmnet().}

\item{newoffset}{If an offset is used in the fit, then one must be supplied for making predictions (except for type="coefficients" or type="nonzero").}

\item{...}{Other arguments are passed directly to backend (currently glmnet() is used), and if not given, backend API's deault values are used.}
}
\value{
The object returned depends on type.
}
\description{
Make predictions from a fitted AccurateGLM
}
