% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred-vars.R
\docType{class}
\name{PredVars-class}
\alias{PredVars-class}
\title{S4 class for predictors}
\description{
S4 class for predictors
}
\section{Slots}{

\describe{
\item{\code{vars_info}}{A list of class VarInfo. Each element represents one predictor and contains information
of it and how to create its matrix representation from \code{data} slot.}

\item{\code{data}}{A data.frame which contains numerical values for creation of matrix representation of predictors.}
}}

