% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-aglm.R
\name{print.AccurateGLM}
\alias{print.AccurateGLM}
\title{Display textual information of the model}
\usage{
\method{print}{AccurateGLM}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{A model object obtained from \code{aglm()} or \code{cv.aglm()}.}

\item{digits}{Used to control significant digits in printout.}

\item{...}{Other arguments are passed directly to \code{print.glmnet()}.}
}
\value{
No return value, called for side effects.
}
\description{
Display textual information of the model
}
\author{
Kenji Kondo
}
